requirements.txt
altgraph==0.17.3
pefile==2023.2.7
Pillow==9.5.0
pyinstaller-hooks-contrib==2023.2
pywin32-ctypes==0.2.0
